feat(uploads): ship ticket flow only; defer server-side URL fetch
Integrate the base64-free upload path from #34, scoped to the ticket flow
(create-upload-ticket / get-upload-result + the /upload/:ticket route) and
hold back upload-file-from-url.

- Remove upload-file-from-url, the SSRF URL fetcher (fetch-url.ts), the host
  allow-list config (LEXWARE_UPLOAD_ALLOWED_HOSTS) and their tests. Keep the
  shared sanitizeFilename helper as src/uploads/filename.ts.
- Mount the /upload routes only when the drafts capability is enabled, so a
  read-only deployment never exposes the unauthenticated write route.
- Serve the ticket page Cache-Control: no-store + X-Content-Type-Options: nosniff.
- Bump to 0.1.11; CHANGELOG records the deferral and its reason.

The URL fetcher carries a DNS-rebinding TOCTOU that is moot for the built-in
Microsoft defaults but live for any custom allow-list; it will be reconsidered
separately with connection-level IP pinning, disabled by default.
