+++ .env.example
@@ -84,14 +84,6 @@ MCP_AUTH_TOKEN=
 # Finalize / legally-binding write tools (issue invoices). IRREVERSIBLE. Default: false
 # LEXWARE_ENABLE_FINALIZE=false
 
-# Hosts that upload-file-from-url may download from (comma-separated). Matched on the host
-# itself or any subdomain, on a dot boundary — "evilsharepoint.com" does NOT match
-# "sharepoint.com". Setting this REPLACES the defaults rather than adding to them, so you can
-# opt out of the Microsoft file-sharing domains entirely. Setting it to an empty value blocks
-# every host, which disables the tool — an empty list never means "allow everything".
-# Default: sharepoint.com,onedrive.live.com,1drv.ms,graph.microsoft.com
-# LEXWARE_UPLOAD_ALLOWED_HOSTS=files.example.com,storage.example.org
-
 # ---- Server ----
 # Listen port. Cloud Run injects this automatically. Default: 8080
 # PORT=8080
